Removing unused code On branch BUGLocationAPI Changes to be committed: -	new file:   config/elasticClient.js -	renamed:    redis.js -> config/redis.js -	modified:   controllers/LocationController.js -	modified:   routes/location.js
